삭제된 게시글 체크 로직 변경

+++ src/main/java/camel/BoostMarketer/blog/service/BlogBatchService.java
@@ -100,8 +100,11 @@ public class BlogBatchService {
             if(!naverPostNoList.isEmpty()){
                 List<String> dbPostNoList = blogMapper.selectPostNoByBlogId(lastPostNoDto.getBlogId());
                 dbPostNoList.removeAll(naverPostNoList);
-                blogMapper.deleteBlogPostByPostId(dbPostNoList);
-                keywordMapper.deleteKeywordRankByPostId(dbPostNoList);
+                if(!dbPostNoList.isEmpty()){
+                    blogMapper.deleteBlogPostByPostId(dbPostNoList);
+                    keywordMapper.deleteKeywordRankByPostId(dbPostNoList);
+                }
+
             }
         }
     }
