+++ src/main/java/camel/BoostMarketer/blog/service/BlogService.java
@@ -290,8 +290,10 @@ public class BlogService {
             if(!naverPostNoList.isEmpty()){
                 List<String> dbPostNoList = blogMapper.selectPostNoByBlogId(lastPostNoDto.getBlogId());
                 dbPostNoList.removeAll(naverPostNoList);
-                blogMapper.deleteBlogPostByPostId(dbPostNoList);
-                keywordMapper.deleteKeywordRankByPostId(dbPostNoList);
+                if(!dbPostNoList.isEmpty()){
+                    blogMapper.deleteBlogPostByPostId(dbPostNoList);
+                    keywordMapper.deleteKeywordRankByPostId(dbPostNoList);
+                }
             }
         }
     }
